requirements.txt
Django==5.1.2
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-decouple==3.8
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
